+++ crates/zng-view/src/lib.rs
@@ -11,8 +11,8 @@
 //!
 //! ```toml
 //! [dependencies]
-//! zng = "0.13.11"
-//! zng-view = "0.7.5"
+//! zng = "0.13.12"
+//! zng-view = "0.7.6"
 //! ```
 //!
 //! Then call `zng::env::init` before any other code in `main` to setup a view-process that uses